Fixed Blocking Error
When both the Client and Server programs are running on the same machine, and if the Client is already connected with another Server, when the Server on the same machine scans the network, it will connect to the already bound Client, and it will wait forever for that Client to send the first bytes, which will never happen.

By adding a time out to TextSocket, this issue has been resolved.