Fixed Timeout Issue
After waiting for 5 seconds for a response from the Client, if such a response is recieved, the TextSocket timeout time is now set from 5 seconds to infinity.

This will ensure that the TextiSocket will not adruptly shutdown if no message is sent from the Client to Server within 5 seconds, which can happen at any time.